Feat : Storyboard 삭제 및 초기 화면 설정
