+++ CloudNotes/CloudNotes/ViewController/SplitViewController.swift
@@ -1,0 +1,15 @@
+//
+//  MainSplitViewController.swift
+//  CloudNotes
+//
+//  Created by 이영우 on 2021/06/02.
+//
+
+import UIKit
+
+class SplitViewController: UISplitViewController {
+  override func viewDidLoad() {
+    super.viewDidLoad()
+    
+  }
+}
